requirements.txt
discord.py
pyyaml
apscheduler
python-dotenv